requirements.txt
streamlit>=1.45.0
tavily-python>=0.5.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
lxml>=5.0.0
python-pptx>=1.0.0
